+++ registry/registry.go
@@ -21,9 +21,10 @@ var (
 
 // Registry holds objects and settings.
 type Registry struct {
-	Config *Configuration
-
+	Config  Configuration
 	Objects []Object
+	// objIDVerMap map contains references to objects stored in Objects slice.
+	objIDVerMap map[int32]map[string]*Object // First key - ObjectID, second - Version
 }
 
 // TODO implement tests for registry
@@ -36,23 +37,21 @@ type Registry struct {
 // Find
 
 // New creates a new registry, using provided or default configuration.
-func New(cfg *Configuration) (*Registry, error) {
-	var err error
-
-	reg := &Registry{
+func New(cfg Configuration) (Registry, error) {
+	reg := Registry{
 		Config:  cfg,
 		Objects: nil,
 	}
 
-	if reg.Config == nil {
-		reg.Config = DefaultConfiguration()
-	}
-
 	if reg.Config.InitOnNew {
+		var err error
+
 		reg.Objects, err = reg.ImportFromAPI()
 		if err != nil {
-			return nil, err
+			return reg, err
 		}
+
+		reg.objIDVerMap = objToMap(reg.Objects)
 	}
 
 	if reg.Config.Sanitize {
@@ -60,6 +59,21 @@ func New(cfg *Configuration) (*Registry, error) {
 	}
 
 	return reg, nil
+}
+
+func objToMap(objects []Object) map[int32]map[string]*Object {
+	objMap := make(map[int32]map[string]*Object)
+
+	for i, object := range objects {
+		_, ok := objMap[object.ObjectID]
+		if !ok {
+			objMap[object.ObjectID] = make(map[string]*Object)
+		}
+
+		objMap[object.ObjectID][object.ObjectVersion] = &objects[i]
+	}
+
+	return objMap
 }
 
 // TODO implement sanitization using regular expressions
@@ -107,7 +121,7 @@ func (r *Registry) Export(filename string) error {
 }
 
 // Import loads objects and resources from file.
-// Overwrites current registry objects and rObesources.
+// Overwrites current registry objects and resources.
 func (r *Registry) Import(filename string) error {
 	if filename == "" {
 		return errEmptyFilename
@@ -118,13 +132,19 @@ func (r *Registry) Import(filename string) error {
 		return err
 	}
 
-	return yaml.Unmarshal(data, &r.Objects)
+	if err := yaml.Unmarshal(data, &r.Objects); err != nil {
+		return err
+	}
+
+	r.objIDVerMap = objToMap(r.Objects)
+
+	return nil
 }
 
 // ImportFromAPI initializes the registry from official OMA API.
 // Overwrites current registry objects and resources.
 // TODO make import asynchronous, run it in separate go routine
-// TODO block Find and Export operations while importing to avoid inconsistent state
+// TODO block Find and Export operations while importing to avoid inconsistent state.
 func (r *Registry) ImportFromAPI() ([]Object, error) {
 	objectsMeta, err := r.getObjectsMeta()
 	if err != nil {
@@ -143,7 +163,15 @@ func (r *Registry) ImportFromAPI() ([]Object, error) {
 			return nil, err
 		}
 
-		objects = append(objects, *object) // TODO nil check
+		if object.ObjectVersion == "" {
+			object.ObjectVersion = DefaultObjectVersion
+		}
+
+		if object.LWM2MVersion == "" {
+			object.LWM2MVersion = DefaultLwM2MVersion
+		}
+
+		objects = append(objects, object)
 	}
 
 	return objects, nil
@@ -151,13 +179,13 @@ func (r *Registry) ImportFromAPI() ([]Object, error) {
 
 // Compare makes comparison of r and reg registries.
 // Returns a list of non-equal objects with difference explanation.
-func (r *Registry) Compare(reg *Registry) []ObjectComparison {
+func (r *Registry) Compare(reg Registry) []ObjectComparison {
 	// TODO store objects in registry in maps to improve lookup performance
 	var objComp []ObjectComparison
 
 	// Compare r with reg
 	for _, regObj := range reg.Objects {
-		if rObj, err := r.Find(&regObj); err != nil {
+		if rObj, err := r.Find(regObj); err != nil {
 			regObjCopy := regObj
 			// rObjCopy := rObj
 
@@ -180,7 +208,7 @@ func (r *Registry) Compare(reg *Registry) []ObjectComparison {
 
 	// Compare reg with r
 	for _, rObj := range r.Objects {
-		if regObj, err := reg.Find(&rObj); err != nil {
+		if regObj, err := reg.Find(rObj); err != nil {
 			// regObjCopy := regObj
 			rObjCopy := rObj
 
@@ -208,10 +236,10 @@ func (r *Registry) Compare(reg *Registry) []ObjectComparison {
 
 // Find looks for an object in current registry.
 // Returns an empty object and error, when object not found.
-func (r *Registry) Find(o *Object) (Object, error) {
-	for _, rObj := range r.Objects {
-		if rObj.ObjectID == o.ObjectID && rObj.ObjectVersion == o.ObjectVersion {
-			return rObj, nil
+func (r *Registry) Find(o Object) (Object, error) {
+	if objVerMap, ok := r.objIDVerMap[o.ObjectID]; ok {
+		if obj, ok := objVerMap[o.ObjectVersion]; ok {
+			return *obj, nil
 		}
 	}
 
@@ -224,13 +252,11 @@ func (r *Registry) Find(o *Object) (Object, error) {
 func (r *Registry) FindObjectsByID(id int32) ([]Object, error) {
 	var objects []Object
 
-	for _, rObj := range r.Objects {
-		if rObj.ObjectID == id {
-			objects = append(objects, rObj)
+	if objVerMap, ok := r.objIDVerMap[id]; ok {
+		for _, object := range objVerMap {
+			objects = append(objects, *object)
 		}
-	}
-
-	if len(objects) == 0 {
+	} else {
 		return nil, errObjNotFound
 	}
 
@@ -281,42 +307,11 @@ func (r *Registry) FindResourcesByObjResIDs(objID, resID int32) ([]Resource, err
 	return resources, nil
 }
 
-// ObjectMeta struct for ObjectMeta
-type ObjectMeta struct {
-	ObjectID          int32  `json:"ObjectID" xml:"ObjectID"`
-	Ver               string `json:"Ver,omitempty" xml:"Ver"`
-	URN               string `json:"URN,omitempty" xml:"URN"`
-	Name              string `json:"Name" xml:"Name"`
-	Description       string `json:"Description,omitempty" xml:"Description"`
-	Owner             string `json:"Owner,omitempty" xml:"Owner"`
-	Label             string `json:"Label,omitempty" xml:"Label"`
-	ObjectLink        string `json:"ObjectLink" xml:"ObjectLink"`
-	ObjectLinkVisible string `json:"ObjectLinkVisible,omitempty" xml:"ObjectLinkVisible"`
-	SpecLink          string `json:"SpecLink,omitempty" xml:"SpecLink"`
-	SpecLinkVisible   string `json:"SpecLinkVisible,omitempty" xml:"SpecLinkVisible"`
-	VortoLink         string `json:"VortoLink,omitempty" xml:"VortoLink"`
-}
+// TODO func (r *Registry) FindResourcesByObjResIDsVer(objID, objVer, resID int32) ([]Resource, error) {
 
 // getObjectsMeta retrieve all objects metadata.
 func (r *Registry) getObjectsMeta() ([]ObjectMeta, error) {
-	u := "http://www.openmobilealliance.org/api/lwm2m/v1/Object"
-
-	client := http.Client{
-		Timeout: 5 * time.Second,
-	}
-
-	resp, err := client.Get(u)
-	if err != nil {
-		return nil, err
-	}
-
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			// TODO log error
-		}
-	}()
-
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := getURL("http://www.openmobilealliance.org/api/lwm2m/v1/Object")
 	if err != nil {
 		return nil, err
 	}
@@ -331,12 +326,27 @@ func (r *Registry) getObjectsMeta() ([]ObjectMeta, error) {
 }
 
 // getObject fetch object details based on metadata.
-func (r *Registry) getObject(objectMeta ObjectMeta) (*Object, error) {
+func (r *Registry) getObject(objectMeta ObjectMeta) (Object, error) {
+	body, err := getURL(objectMeta.ObjectLink)
+	if err != nil {
+		return Object{}, err
+	}
+
+	var lwm2m Lwm2M
+
+	if err := xml.Unmarshal(body, &lwm2m); err != nil {
+		return Object{}, err
+	}
+
+	return lwm2m.Object, nil
+}
+
+func getURL(url string) ([]byte, error) {
 	client := http.Client{
 		Timeout: 5 * time.Second,
 	}
 
-	resp, err := client.Get(objectMeta.ObjectLink)
+	resp, err := client.Get(url)
 	if err != nil {
 		return nil, err
 	}
@@ -347,16 +357,5 @@ func (r *Registry) getObject(objectMeta ObjectMeta) (*Object, error) {
 		}
 	}()
 
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	var lwm2m Lwm2M
-
-	if err := xml.Unmarshal(body, &lwm2m); err != nil {
-		return nil, err
-	}
-
-	return &lwm2m.Object, nil
+	return ioutil.ReadAll(resp.Body)
 }